+++ node/resolvers/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 /* eslint-disable @typescript-eslint/naming-convention */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { ForbiddenError } from '@vtex/api'
@@ -145,6 +144,69 @@ const MUTATIONS = {
 
 export const resolvers = {
   Routes: {
+    checkout: async (ctx: Context) => {
+      const {
+        vtex: { storeUserAuthToken, sessionToken, logger },
+        clients: { session, masterdata },
+      } = ctx
+
+      const token: any = storeUserAuthToken
+      const response: any = {}
+
+      ctx.response.status = !token ? 403 : 200
+
+      if (token) {
+        const sessionData = await session
+          .getSession(sessionToken as string, ['*'])
+          .then((currentSession: any) => {
+            return currentSession.sessionData
+          })
+          .catch((error: any) => {
+            logger.error({
+              message: 'getSession-error',
+              error,
+            })
+
+            return null
+          })
+
+        if (sessionData?.namespaces['storefront-permissions']) {
+          if (
+            sessionData.namespaces['storefront-permissions']?.organization
+              ?.value
+          ) {
+            const organization = await masterdata.getDocument({
+              dataEntity: ORGANIZATION_DATA_ENTITY,
+              fields: ['paymentTerms'],
+              id:
+                sessionData.namespaces['storefront-permissions']?.organization
+                  ?.value,
+            })
+
+            response.organization = organization
+          }
+
+          if (
+            sessionData.namespaces['storefront-permissions']?.costcenter?.value
+          ) {
+            const costcenter = await masterdata.getDocument({
+              dataEntity: COST_CENTER_DATA_ENTITY,
+              fields: ['addresses'],
+              id:
+                sessionData.namespaces['storefront-permissions']?.costcenter
+                  ?.value,
+            })
+
+            response.costcenter = costcenter
+          }
+        }
+      }
+
+      ctx.set('Content-Type', 'application/json')
+      ctx.set('Cache-Control', 'no-cache, no-store')
+
+      ctx.response.body = response
+    },
     orders: async (ctx: Context) => {
       const {
         vtex: { storeUserAuthToken, sessionToken, logger },
@@ -164,8 +226,11 @@ export const resolvers = {
         .then((currentSession: any) => {
           return currentSession.sessionData
         })
-        .catch((err: any) => {
-          logger.error(err)
+        .catch((error: any) => {
+          logger.error({
+            message: 'getSession-error',
+            error,
+          })
 
           return null
         })
@@ -199,9 +264,11 @@ export const resolvers = {
             },
           }
         )
-        .catch((err: any) => {
-          console.log('Error quering permissions =>', err)
-          logger.error(err)
+        .catch((error: any) => {
+          logger.error({
+            message: 'checkUserPermission-error',
+            error,
+          })
 
           return {
             data: {
@@ -819,7 +886,6 @@ export const resolvers = {
           return result.data.saveUser
         })
         .catch((error: any) => {
-          console.error(error)
           logger.error({
             message: 'addUser-error',
             error,
@@ -872,7 +938,6 @@ export const resolvers = {
           return result.data.deleteUser
         })
         .catch((error: any) => {
-          console.error(error)
           logger.error({
             message: 'deleteUser-error',
             error,