+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `/b2b/oms/user/checkout/` route to return `paymentTerms` and Cost Center addresses
+
 ## [0.3.2] - 2021-11-03
 
 ### Changed